:recycle:  moving some components from common.tsx to separate files
This is direct copy & paste with no changes. Just moving things a bit to reduce the size of the
massive common.tsx file
