+++ upcoming-release-notes/1248.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [MatissJanis]
+---
+
+Moving some components from `common.tsx` to separate files inside the `common` folder
